+++ impl/src/main/java/io/kroxylicious/testing/kafka/common/BrokerConfig.java
@@ -18,15 +18,11 @@ import io.kroxylicious.testing.kafka.api.KafkaClusterConstraint;
  * {@code @BrokerConfig} can be used to annotate a field in a test class or a
  *  parameter in a lifecycle method or test method of type {@link KafkaCluster}
  *  constraining the cluster to have the given broker configuration.
- *
- * @deprecated use {@link KafkaConfig} instead.  If this BrokerConfig and KafkaConfig are applied together and
- * both provide the same name, the value provided by KafkaConfig takes precedence.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.FIELD, ElementType.PARAMETER })
 @Repeatable(BrokerConfig.List.class)
 @KafkaClusterConstraint
-@Deprecated(forRemoval = true, since = "0.7.0")
 public @interface BrokerConfig {
     /**
      * The name of the <a href="https://kafka.apache.org/documentation.html#brokerconfigs">broker configuration parameter</a>.